finished real_escaping strings in admin panel

+++ admin/includes/view_all_comments.php
@@ -6,7 +6,7 @@ if (isset($_POST['checkBoxArray'])){
 
     foreach ($_POST['checkBoxArray'] as $comment_value_id){
 
-        $bulk_options = $_POST['bulk_options'];
+        $bulk_options = escape($_POST['bulk_options']);
 
         switch ($bulk_options) {
 
@@ -85,13 +85,13 @@ if (isset($_POST['checkBoxArray'])){
         $select_comments = mysqli_query($connection, $query);
 
         while ($row = mysqli_fetch_assoc($select_comments)) {
-            $comment_id = $row['comment_id'];
-            $comment_post_id = $row['comment_post_id'];
-            $comment_author = $row['comment_author'];
-            $comment_content = $row['comment_content'];
-            $comment_email = $row['comment_email'];
-            $comment_status = $row['comment_status'];
-            $comment_date = $row['comment_date'];
+            $comment_id = escape($row['comment_id']);
+            $comment_post_id = escape($row['comment_post_id']);
+            $comment_author = escape($row['comment_author']);
+            $comment_content = escape($row['comment_content']);
+            $comment_email = escape($row['comment_email']);
+            $comment_status = escape($row['comment_status']);
+            $comment_date = escape($row['comment_date']);
 
 
             echo "<tr>";
@@ -114,8 +114,8 @@ if (isset($_POST['checkBoxArray'])){
 
             while($row = mysqli_fetch_assoc($select_post_id_query)){
 
-                $post_id = $row['post_id'];
-                $post_title = $row['post_title'];
+                $post_id = escape($row['post_id']);
+                $post_title = escape($row['post_title']);
 
                 echo "<td><a href='../post.php?p_id=$post_id'>$post_title</td></td>";
 
@@ -141,33 +141,52 @@ if (isset($_POST['checkBoxArray'])){
 
     if (isset($_GET['approve'])){
 
-        $the_comment_id = $_GET['approve'];
+        if (isset($_SESSION['user_role'])) {
 
-        $query = "UPDATE comments SET comment_status = 'approved' WHERE comment_id = $the_comment_id ";
-        $approve_query = mysqli_query($connection, $query);
+            if ($_SESSION['user_role'] == 'admin') {
 
-        header("Location: comments.php");
+
+                $the_comment_id = $_GET['approve'];
+
+                $query = "UPDATE comments SET comment_status = 'approved' WHERE comment_id = $the_comment_id ";
+                $approve_query = mysqli_query($connection, $query);
+
+                header("Location: comments.php");
+            }
+        }
     }
 
 
     if (isset($_GET['unapprove'])){
 
-        $the_comment_id = $_GET['unapprove'];
+        if (isset($_SESSION['user_role'])) {
 
-        $query = "UPDATE comments SET comment_status = 'unapproved' WHERE comment_id = $the_comment_id ";
-        $unapprove_query = mysqli_query($connection, $query);
+            if ($_SESSION['user_role'] == 'admin') {
 
-        header("Location: comments.php");
+                $the_comment_id = escape($_GET['unapprove']);
+
+                $query = "UPDATE comments SET comment_status = 'unapproved' WHERE comment_id = $the_comment_id ";
+                $unapprove_query = mysqli_query($connection, $query);
+
+                header("Location: comments.php");
+            }
+        }
     }
 
 if (isset($_GET['delete'])){
 
-    $the_comment_id = $_GET['delete'];
+    if (isset($_SESSION['user_role'])) {
 
-    $query = "DELETE FROM comments WHERE comment_id = {$the_comment_id}";
-    $delete_query = mysqli_query($connection, $query);
+        if ($_SESSION['user_role'] == 'admin') {
 
-    header("Location: comments.php");
+            $the_comment_id = escape($_GET['delete']);
+
+            $query = "DELETE FROM comments WHERE comment_id = {$the_comment_id}";
+            $delete_query = mysqli_query($connection, $query);
+
+            header("Location: comments.php");
+        }
+    }
 }
 
 ?>
